+++ src/server/DurationAdapter.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
