+++ src/server/LocalDateTimeAdapter.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 import com.google.gson.*;
 import java.lang.reflect.Type;
