+++ src/server/PrioritizedTasksHandler.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
