+++ src/server/SubtasksHandler.java
@@ -1,75 +1,63 @@
-package Server;
+package server;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import exceptions.NotFoundException;
-import tasks.Epic;
 import tasks.Subtask;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 
-public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
+public class SubtasksHandler extends BaseHttpHandler implements HttpHandler {
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         String method = exchange.getRequestMethod();
         String path = exchange.getRequestURI().getPath();
         String[] arrayPath = path.split("/");
         boolean isHaveId = false;
-        boolean isHaveSubtask = false;
         int id = 0;
         if (arrayPath.length > 2) {
             isHaveId = true;
             id = Integer.parseInt(arrayPath[2]);
-            if (arrayPath.length > 3) {
-                isHaveSubtask = true;
-            }
         }
         switch (method) {
             case "GET":
                 if (isHaveId) {
-                    Epic epic;
+                    Subtask subtask;
                     try {
-                        epic = taskManager.getEpic(id);
+                        subtask = taskManager.getSubtask(id);
                     } catch (NotFoundException e) {
                         sendNotFound(exchange, e.getMessage());
                         break;
                     }
-                    if (isHaveSubtask) {
-                        List<Subtask> subtasksOfEpic;
-                        try {
-                            subtasksOfEpic = taskManager.getSubtasksOfEpic(id);
-                        } catch (NotFoundException e) {
-                            sendNotFound(exchange, e.getMessage());
-                            break;
-                        }
-                        String json = gson.toJson(subtasksOfEpic);
-                        sendText(exchange, json);
-                        break;
-                    }
-                    String json = gson.toJson(epic);
+                    String json = gson.toJson(subtask);
                     sendText(exchange, json);
                 } else {
-                    List<Epic> listOfEpics = taskManager.getListEpic();
-                    String json = gson.toJson(listOfEpics);
+                    List<Subtask> listOfSubtasks = taskManager.getListSubtask();
+                    String json = gson.toJson(listOfSubtasks);
                     sendText(exchange, json);
                 }
                 break;
             case "POST":
                 InputStream inputStream = exchange.getRequestBody();
-                String jsonEpic = new String(inputStream.readAllBytes(), getUtf());
-                Epic epic = gson.fromJson(jsonEpic, Epic.class);
+                String jsonSubtask = new String(inputStream.readAllBytes(), getUtf());
+                Subtask subtask = gson.fromJson(jsonSubtask, Subtask.class);
+                boolean isCross = false;
+                Integer idCross = 0;
                 if (isHaveId) {
-                    taskManager.updateEpic(epic);
+                    isCross = taskManager.updateSubtask(subtask); // Если задача пересекается, то updateTask возвращает true
                 } else {
-                    taskManager.addEpic(epic);
+                    idCross = taskManager.addSubtask(subtask); // Если задача пересекается, то addTask возвращает Integer = null
+                }
+                if (isCross || idCross == null) {
+                    sendHasInteractions(exchange, "Subtask is crossing the other tasks");
                 }
                 sendWithoutAnswer(exchange);
                 break;
             case "DELETE":
                 if (isHaveId) {
-                    taskManager.removeEpic(id);
-                    sendText(exchange, "Epic is deleted");
+                    taskManager.removeSubtask(id);
+                    sendText(exchange, "Subtask is deleted");
                 } else {
                     sendNotFound(exchange, "ID is not send");
                 }
